AT19_MACHINE_LEARNING/requirements.txt
python-dotenv==1.0.0
Flask==2.2.3
Flask-RESTful==0.3.9
Flask-Cors==3.0.10
zope.interface==5.5.2
pymongo==4.3.3
flask-swagger-ui==4.11.1
imageai==3.0.3
Cython==0.29.33
Pillow==9.4.0
torch==1.13.1+cpu
torchvision==0.14.1+cpu
pytest==7.1.3
tqdm==4.64.1
scipy>=1.7.3
mock==4.0.3
matplotlib>=3.4.3
opencv-python==4.7.0.72
deepface==0.0.79
face-recognition==1.3.0
dlib==19.24.0

